requirements.txt
fastapi==0.116.2
uvicorn==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
pyjwt==2.10.1
bcrypt==4.3.0
passlib==-1.7.4
SQLAlchemy==2.0.41
asyncpg==0.30.0
aiosqlite==0.20.0
dnspython==2.8.0
email-validator==2.3.0
curl_cffi==0.7.1
pydantic-settings==2.9.1

